requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
asyncpg==0.30.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg==3.2.3
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
